Module auth moved to application folder
