+++ src/router/AppRouter.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Routes, Route, BrowserRouter, Navigate } from "react-router-dom";
 
 import { Home } from "../application/home";
-import { LoginPage } from "../auth";
+import { LoginPage } from "../application/auth";
 
 export const AppRouter = () => {
   return (
